==> analysis/01-preprocessing.Rmd <==
---
title: "Pre-processing"
---

```{r knitr, include = FALSE}
DOCNAME = "01-preprocessing"
NOW <- Sys.time()

# Time chunks during knitting
knitr::knit_hooks$set(timeit = function(before) {
    if (before) {
        print(paste("Start:", Sys.time()))
        NOW <<- Sys.time()
    } else {
        print(paste("Stop:", Sys.time()))
        print(Sys.time() - NOW)
    }
})

knitr::opts_chunk$set(
    autodep        = TRUE,
    cache          = TRUE,
    cache.path     = paste0("cache/", DOCNAME, "/"),
    cache.comments = FALSE,
    echo           = TRUE,
    error          = FALSE,
    fig.align      = "center",
    fig.width      = 10,
    fig.height     = 8,
    message        = FALSE,
    warning        = FALSE,
    timeit         = TRUE
)
```

```{r libaries, cache = FALSE}
# scRNA-seq
library("DropletUtils")
library("SingleCellExperiment")

# Plotting
library("cowplot")
library("UpSetR")
library("grid")

# Tidyverse
library("tidyverse")
```

```{r source, cache = FALSE}
source(here::here("R/load.R"))
source(here::here("R/annotate.R"))
source(here::here("R/output.R"))
```

```{r bpparam, cache = FALSE}
bpparam <- BiocParallel::MulticoreParam(workers = 10)
```

Introduction
============

In this document we are going to read in the unfiltered counts matrix produced
by `Cell Ranger` and determine which of those droplets contain cells using the
`DropletUtils` package.

```{r load}
path <- here::here("data/cellranger")
raw <- read10x(path, dataset = "Orgs123")

filt_barcodes <- readr::read_lines(file.path(path, "filtered_barcodes.tsv.gz"))

colData(raw)$CellRangerFilt <- paste(colData(raw)$Barcode, 
                                     colData(raw)$Sample, sep = "-") %in%
    filt_barcodes
```

The raw dataset has `r nrow(raw)` features and `r ncol(raw)` droplets.

```{r all-zero}
raw <- raw[Matrix::rowSums(counts(raw)) > 0, Matrix::colSums(counts(raw)) > 0]
```

After removing all zero features and droplets the dataset has `r nrow(raw)` 
features and `r ncol(raw)` droplets.

Barcode ranks
=============

```{r barcodes}
empty_thresh <- 100

bc_ranks <- barcodeRanks(counts(raw), lower = empty_thresh)

colData(raw)$BarcodeRank   <- bc_ranks$rank
colData(raw)$BarcodeTotal  <- bc_ranks$total
colData(raw)$BarcodeFitted <- bc_ranks$fitted

bc_data <- colData(raw) %>%
    as.data.frame() %>%
    select(Cell, Kept = CellRangerFilt, Rank = BarcodeRank,
           Total = BarcodeTotal, Fitted = BarcodeFitted) %>%
    arrange(Rank)
```

Let's start by ordering the droplets according to their total counts and
plotting this on a log scale. This let's us see the distribution of total
counts.

```{r barcodes-plot}
ggplot(bc_data, aes(x = Rank, y = Total)) +
    geom_point(shape = 1, aes(colour = Kept)) +
    geom_line(aes(y = Fitted), colour = "red") +
    geom_hline(yintercept = bc_ranks$knee,
               colour = "dodgerblue", linetype = "dashed") +
    annotate("text", x = 0, y = bc_ranks$knee, label = "Knee",
             colour = "dodgerblue", hjust = 0, vjust = -1) +
    geom_hline(yintercept = bc_ranks$inflection,
               colour = "forestgreen", linetype = "dashed") +
    annotate("text", x = 0, y = bc_ranks$inflection, label = "Inflection",
             colour = "forestgreen", hjust = 0, vjust = -1) +
    geom_hline(yintercept = empty_thresh,
               colour = "darkorchid", linetype = "dashed") +
    annotate("text", x = 0, y = empty_thresh, label = "Empty threshold",
             colour = "darkorchid", hjust = 0, vjust = -1) +
    scale_x_log10(labels = scales::number) +
    scale_y_log10(labels = scales::number) +
    scale_colour_manual(values = c("black", "violet")) +
    ylab("Total counts") +
    theme_minimal()
```

This is typical of what we see for 10x experiment where there is a sharp drop
off between droplets with lots of counts and those without many. The inflection
and knee points are methods for identifying the transition between
distributions. These are roughly associated with the cells selected by `Cell
Ranger` The empty threshold line indicates the point at which we assume droplets
must be empty (total counts <= `r empty_thresh`).

Traditional Cell Ranger
=======================

The version of Cell Ranger we have used here selects cells using a modified
version of the `EmptyDrops` method (see below) but older versions used an
alternative approach by calculating the 99th percentile of the total number
of counts in top given expected number of cells and selecting droplets that
had at least 10 percent of this many counts. For comparison purposes we will
perform this method as well.

```{r default-drops}
default1 <- defaultDrops(counts(raw[, colData(raw)$Sample == 1]))
default2 <- defaultDrops(counts(raw[, colData(raw)$Sample == 2]))
default3 <- defaultDrops(counts(raw[, colData(raw)$Sample == 3]))
colData(raw)$DefaultFilt <- c(default1, default2, default3)

plot_data <- colData(raw) %>%
    as.data.frame() %>%
    select(Cell, Sample, Total = BarcodeTotal, Kept = DefaultFilt) %>%
    group_by(Sample) %>%
    mutate(Rank = rank(-Total)) %>%
    arrange(Sample, Rank)

ggplot(plot_data, aes(x = Rank, y = Total)) +
    geom_point(shape = 1, aes(colour = Kept)) +
    scale_x_log10(labels = scales::number) +
    scale_y_log10(labels = scales::number) +
    scale_colour_manual(values = c("black", "forestgreen")) +
    facet_wrap(~ Sample, nrow = 1, labeller = label_both) +
    ylab("Total counts") +
    theme_minimal()
```

This method identifies `r sum(colData(raw)$DefaultFilt)` droplets as containing
cells.

Empty drops
===========

We will now look at identifying which droplets to select using the `EmptyDrops`
method. This method tests whether the composition of a droplet is significantly
different from the ambient RNA in the sample which is obtained by pooling the
empty droplets. Droplets with very large counts are also automatically retained.

```{r empty-drops}
set.seed(1)

emp_iters <- 30000

emp_drops <- emptyDrops(counts(raw), lower = empty_thresh, niters = emp_iters,
                        test.ambient = TRUE, BPPARAM = bpparam)
```

`EmptyDrops` calculates p-values using a permutation approach. Let's check that
we are usually a sufficient number of iterations. If there are any droplets
that have non-significant p-values but are limited by the number of permuations
the number should be increased.

```{r empty-drops-iters}
emp_fdr <- 0.01
is_cell <- emp_drops$FDR <= emp_fdr
is_cell[is.na(is_cell)] <- FALSE

colData(raw)$EmpDropsLogProb <- emp_drops$LogProb
colData(raw)$EmpDropsPValue  <- emp_drops$PValue
colData(raw)$EmpDropsLimited <- emp_drops$Limited
colData(raw)$EmpDropsFDR     <- emp_drops$FDR
colData(raw)$EmpDropsFilt    <- is_cell

table(Limited = emp_drops$Limited, Significant = is_cell)
```

Another way to check the `EmptyDrops` results is to look at the droplets below
our empty threshold. We are assuming that these droplets only contain ambient
RNA and therefore the null hypothesis should be true and the distribution of
p-values should be approximately uniform.

```{r empty-drops-pvals}
colData(raw) %>%
    as.data.frame() %>%
    filter(BarcodeTotal <= empty_thresh,
           BarcodeTotal > 0) %>%
    ggplot(aes(x = EmpDropsPValue)) +
    geom_histogram() +
    xlab("p-value") +
    theme_minimal()
```

Peaks near zero would tell us that not all of the droplets below the threshold
are truly empty and that we should lower it. 

We can also plot the negative log-probability against the total counts to see
which droplets `EmptyDrops` has selected.

```{r empty-drops-plot}
colData(raw) %>%
    as.data.frame() %>%
    filter(!is.na(EmpDropsFilt)) %>%
    ggplot(aes(x = BarcodeTotal, y = -EmpDropsLogProb, colour = EmpDropsFilt)) +
    geom_point() +
    scale_colour_discrete(name = "Significant") +
    xlab("Total counts") +
    ylab("-log(probability)") +
    theme_minimal()
```

There were `r sum(emp_drops$Total <= empty_thresh)` droplets with less than
`r empty_thresh` counts which were used to make up the ambient RNA pool. Of the
remaining `r sum(emp_drops$Total > empty_thresh)` droplets `r sum(is_cell)` were
found to have profiles significantly different from the ambient RNA and should
contain cells.

Comparison
==========

Let's quickly compare differences between the selection methods using an UpSet
plot:

```{r compare-overlap}
plot_data <- list(
    "Cell Ranger"    = colData(raw)$Cell[colData(raw)$DefaultFilt],
    "Cell Ranger v3" = colData(raw)$Cell[colData(raw)$CellRangerFilt],
    "EmptyDrops"     = colData(raw)$Cell[colData(raw)$EmpDropsFilt]
)

upset(fromList(plot_data), order.by = "freq",
      sets.x.label = "Number of cells", text.scale = c(2, 1.2, 2, 1.2, 2, 3))
```

We can see that most of the cells are identified by all methods. A large number
of cells are also identified by the new Cell Ranger and `EmptyDrops` methods.
These are likely to be those cells that fall below the total counts threshold
selected by the traditional Cell Ranger algorithm. Our use of the `EmptyDrops`
algorithm has identified even more cells than Cell Ranger v3 but there are
very few cells that Cell Ranger identifed which `EmptyDrops` didn't.

Selection
=========

We are going to perform further quality control of these cells anyway so at
this stage we will keep those that were selected by either Cell Ranger v3 or
`EmptyDrops`.

```{r select}
selected <- raw[, colData(raw)$CellRangerFilt | colData(raw)$EmpDropsFilt]
selected <- selected[Matrix::rowSums(counts(selected)) > 0, ]

colData(selected)$SelMethod <- "Both"
colData(selected)$SelMethod[!colData(selected)$CellRangerFilt] <- "emptyDrops"
colData(selected)$SelMethod[!colData(selected)$EmpDropsFilt] <- "CellRanger"
```

Annotation
==========

Now that we have a dataset that just contains actual cells we will add some
extra annotation. This includes downloading feature annotation from `BioMart`
and assigning cell cycle stages using the `cyclone` function in the `scran`
package as well as calculating a range of quality control metrics using
`scater`.

```{r annotate}
selected <- annotateSCE(selected,
                        org        = "human",
                        add_anno   = TRUE,
                        host       = "jul2018.archive.ensembl.org",
                        calc_qc    = TRUE,
                        calc_cpm   = TRUE,
                        cell_cycle = TRUE,
                        BPPARAM    = bpparam)

glimpse(as.data.frame(colData(selected)))
glimpse(as.data.frame(rowData(selected)))
```

Figure
======

```{r figure}
plot_data <- colData(raw) %>%
    as.data.frame() %>%
    select(Cell, Kept = EmpDropsFilt, Rank = BarcodeRank,
           Total = BarcodeTotal, Fitted = BarcodeFitted) %>%
    arrange(Rank)

emp_plot <- plot_data %>%
    filter(Total > 0) %>%
    ggplot(aes(x = Rank, y = Total)) +
    geom_point(shape = 1, aes(colour = Kept)) +
    geom_line(aes(y = Fitted), colour = "#00B7C6", size = 1) +
    geom_hline(yintercept = bc_ranks$knee,
               colour = "#00ADEF", linetype = "dashed") +
    annotate("text", x = 0, y = bc_ranks$knee, label = "Knee",
             colour = "#00ADEF", hjust = 0, vjust = -1) +
    geom_hline(yintercept = bc_ranks$inflection,
               colour = "#8DC63F", linetype = "dashed") +
    annotate("text", x = 0, y = bc_ranks$inflection, label = "Inflection",
             colour = "#8DC63F", hjust = 0, vjust = -1) +
    geom_hline(yintercept = empty_thresh,
               colour = "#7A52C7", linetype = "dashed") +
    annotate("text", x = 0, y = empty_thresh, label = "Empty threshold",
             colour = "#7A52C7", hjust = 0, vjust = -1) +
    scale_x_log10(labels = scales::number, breaks = 10 ^ seq(0, 6)) +
    scale_y_log10(labels = scales::number, breaks = 10 ^ seq(0, 5)) +
    scale_colour_manual(values = c("black", "#EC008C"),
                        labels = c("Empty", "Cell")) +
    ggtitle("EmptyDrops droplet selection") +
    ylab("Total counts") +
    theme_minimal() +
    theme(legend.position = "bottom",
          legend.title = element_blank(),
          plot.margin = unit(c(5.5, 20, 5.5, 5.5), "points"))

ggsave(here::here("output", DOCNAME, "droplet-selection.pdf"), emp_plot,
       width = 7, height = 5, scale = 1)
ggsave(here::here("output", DOCNAME, "droplet-selection.png"), emp_plot,
       width = 7, height = 5, scale = 1)

emp_plot
```

```{r fig-comp, collapse = TRUE, fig.show = "hide"}
plot_data <- colData(raw) %>%
    as.data.frame() %>%
    select(Name             = Cell,
           `Cell Ranger`    = DefaultFilt,
           `Cell Ranger v3` = CellRangerFilt,
           EmptyDrops       = `EmpDropsFilt`,
           `Total counts`   = BarcodeTotal) %>%
    mutate(`Cell Ranger`    = if_else(`Cell Ranger`,    1L, 0L),
           `Cell Ranger v3` = if_else(`Cell Ranger v3`, 1L, 0L),
           EmptyDrops       = if_else(EmptyDrops,       1L, 0L)) %>%
    mutate(`Total counts`   = log10(`Total counts`))

upset(plot_data, order.by = "freq",
      sets.x.label = "Number of cells",
      text.scale = c(2, 1.6, 2, 1.3, 2, 3),
      matrix.color = "#7A52C7",
      main.bar.color = "#7A52C7",
      sets.bar.color = "#7A52C7")
grid.edit("arrange", name = "arrange2")
comp_plot <- grid.grab()

ggsave(here::here("output", DOCNAME, "selection-comparison.pdf"), comp_plot,
       width = 7, height = 4, scale = 1.2)
ggsave(here::here("output", DOCNAME, "selection-comparison.png"), comp_plot,
       width = 7, height = 4, scale = 1.2)
```

Summary
=======

Parameters
----------

This table describes parameters used and set in this document.

```{r parameters}
params <- list(
    list(
        Parameter = "n_droplets",
        Value = ncol(raw),
        Description = "Number of droplets in the raw dataset"
    ),
    list(
        Parameter = "empty_thresh",
        Value = empty_thresh,
        Description = "Droplets with less than this many counts are empty"
    ),
    list(
        Parameter = "emp_iters",
        Value = emp_iters,
        Description = "Number of iterations for EmptyDrops p-values"
    ),
    list(
        Parameter = "emp_fdr",
        Value = emp_fdr,
        Description = "FDR cutoff for EmptyDrops"
    ),
    list(
        Parameter = "n_default",
        Value = sum(colData(raw)$DefaultFilt),
        Description = "Number of cells selected by the default Cell Ranger method"
    ),
    list(
        Parameter = "n_cellranger",
        Value = sum(colData(raw)$CellRangerFilt),
        Description = "Number of cells selected by the Cell Ranger v3 method"
    ),
    list(
        Parameter = "n_empdrops",
        Value = sum(colData(raw)$EmpDropsFilt),
        Description = "Number of cells selected by the EmptyDrops method"
    ),
    list(
        Parameter = "n_cells",
        Value = ncol(selected),
        Description = "Number of cells selected"
    )
)

params <- jsonlite::toJSON(params, pretty = TRUE)
knitr::kable(jsonlite::fromJSON(params))
```

Output files
------------

This table describes the output files produced by this document. Right click
and _Save Link As..._ to download the results.

```{r save}
write_rds(selected, here::here("data/processed/01-selected.Rds"))
```

```{r output-barcodes, results = "hide"}
dir.create(here::here("output", DOCNAME), showWarnings = FALSE)

for (sample in unique(colData(selected)$Sample)) {
    barcodes <- colData(selected) %>%
        as.data.frame() %>%
        filter(Sample == sample) %>%
        select(Barcode) %>%
        mutate(Barcode = paste0(Barcode, "-1"))
    
    readr::write_tsv(barcodes,
                     print(here::here("output", DOCNAME,
                                glue::glue("barcodes_Org{sample}.tsv"))),
                     col_names = FALSE)
}
```

```{r output}
readr::write_lines(params, here::here("output", DOCNAME, "parameters.json"))

knitr::kable(data.frame(
    File = c(
        getDownloadLink("parameters.json", DOCNAME),
        getDownloadLink("barcodes_Org1.tsv", DOCNAME),
        getDownloadLink("barcodes_Org2.tsv", DOCNAME),
        getDownloadLink("barcodes_Org3.tsv", DOCNAME),
        getDownloadLink("droplet-selection.png", DOCNAME),
        getDownloadLink("droplet-selection.pdf", DOCNAME),
        getDownloadLink("selection-comparison.png", DOCNAME),
        getDownloadLink("selection-comparison.pdf", DOCNAME)
    ),
    Description = c(
        "Parameters set and used in this analysis",
        "Selected barcodes for organoid 1",
        "Selected barcodes for organoid 2",
        "Selected barcodes for organoid 3",
        "Droplet selection figure (PNG)",
        "Droplet selection figure (PDF)",
        "Selection comparison figure (PNG)",
        "Selection comparison figure (PDF)"
    )
))
```

Session information
-------------------

```{r session-info, cache = FALSE}
devtools::session_info()
```
